# Benchmark vectorized `reverse` and `reverse_copy`, use traits, optimize `reverse_copy` tail